Module srml_support_procedural_tools::syn_ext

Extension to syn types, mainly for parsing

Structs

Braces

Brackets

Meta
Note that syn Meta is almost fine for use case (lacks only ToToken)

NoTrailing

Opt

OuterAttributes

Parens

PunctuatedInner

StopParse
stop parsing here getting remaining token as content Warn duplicate stream (part of)

Trailing

Functions

extract_type_option

has_parametric_type
check if type has a type parameter, defaults to true for some cases.

has_parametric_type_def

is_parametric_type
check if type has any type parameter, defaults to true for some cases.

is_parametric_type_def

Type Definitions

Punctuated

PunctuatedTrailing